+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+* Fix mandatory SERVICE parameters for POST WPS requests
 * Deprecate `host_proxy` in favor of `proxy_url` option
 * Define explicit `HTTP\_PROXY` option
     - Handling of proxyfied urls may be disabled