+++ CHANGELOG.md
@@ -4,7 +4,10 @@
 ## Unreleased
 
 * Fix schemas in ogc api /processes description.
-* Load layer default styles with `loadDefaultStyle`
+* Load layer default styles with `loadDefaultStyle
+
+## 1.8.3 - 2022-12-08
+
 * Fix invalid decoding of complex data as base64
     - Fix https://github.com/3liz/py-qgis-wps/issues/31
 * Fix mandatory SERVICE parameters for POST WPS requests